prizm_server/requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==3.0.6  # https://github.com/un33k/python-slugify
Pillow==6.2.1  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
redis==3.3.11  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==2.2.6  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.40.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.8.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.10.3  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
django-cors-headers==3.1.1
awsebcli==3.15.3
requests==2.20.1
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
python-dateutil==2.8.1
channels==2.3.1
channels-redis==2.4.1
django-imagekit==4.0.2
daphne==2.4.0
jsonfield==2.0.2
django-ckeditor==5.8.0